creating base of requests methods
