+++ lib/tshield/request_matching.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'tshield/request'
+
+module TShield
+  # Class to check request matching
+  class RequestMatching
+  end
+end
